Pend test that fails intermittently
This has been failing for some time and seems very likely to be
related to a change in behavior of rabbitmq -- where we are testing
locally with latest brew installed rabbitmq.

The failure occurs because the first message received in the test is
the one sent in the node spec "hello everybody".